refactor: port arch/IArchStringTests to QtTests
