+++ src/unittests/CMakeLists.txt
@@ -41,6 +41,7 @@ endfunction()
 enable_testing()
 find_package(Qt6 ${REQUIRED_QT_VERSION} REQUIRED COMPONENTS Test)
 
+add_subdirectory(arch)
 add_subdirectory(common)
 add_subdirectory(deskflow)
 add_subdirectory(gui)
